convert apps/odyssey-stats/src/lib/create-odyssey-config.js to ts
