+++ config/production.json
@@ -52,6 +52,7 @@
 		"i18n/translation-scanner": true,
 		"importer/unified": true,
 		"inline-help": true,
+		"is_running_in_jetpack_site": false,
 		"jetpack/agency-dashboard": true,
 		"jetpack/api-cache": false,
 		"jetpack/backup-messaging-i3": true,
